refactor: make ScrollableColumn private
