+++ flutter_news_example/packages/app_ui/lib/src/widgets/widgets.dart
@@ -5,5 +5,4 @@ export 'app_logo.dart';
 export 'app_switch.dart';
 export 'app_text_field.dart';
 export 'content_theme_override_builder.dart';
-export 'scrollable_column.dart';
 export 'show_app_modal.dart';
